pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "box-sandbox"
version = "0.1.0"
description = "Secure, ephemeral sandboxed execution environments — local-first, no Docker required."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
box = "box.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/box"]
